src/main/java/au/com/riosoftware/scart/service/ProductService.java
package au.com.riosoftware.scart.service;

import au.com.riosoftware.scart.model.Product;
import au.com.riosoftware.scart.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {

    private final ProductRepository productRepository;

    @Autowired
    public ProductService(final ProductRepository productRepository) {
        this.productRepository = productRepository;
    }

    public List<Product> getProducts() {
        return this.productRepository.findAll();
    }
}
